move weatherprovider to services.
