--- Website/htdocs/mpmanager/app/Lib/IWeatherDataProvider.php
@@ -1,15 +1,0 @@
-<?php
-
-
-namespace App\Lib;
-
-
-use Psr\Http\Message\ResponseInterface;
-
-interface IWeatherDataProvider
-{
-    public function getCurrentWeatherData(Array $geoLocation): ResponseInterface;
-
-
-    public function getWeatherForeCast(Array $geoLocation): ResponseInterface;
-}
